fix(ci): resolve clang-tidy failures across all CI platforms
- macOS (clang-tidy 22): drop redundant (u64) cast on stat.st_ino in
  app/filesystem_unix.c (ino_t is already u64 on macOS).
- Linux (clang-tidy 18): suppress bugprone-sizeof-expression false positive
  for the pointer-element pages array in core/repository.c.
- Windows (GNU-mode clang defines _MSC_VER on the MSVC ABI target):
  - common.h: only genuine MSVC (cl.exe) uses __forceinline; clang keeps
    static inline __attribute__((always_inline)) so __forceinline (an MS
    extension token) no longer fails under -Wpedantic.
  - atomics.h: gate the _Interlocked* backend on !__clang__ (SK_ATOMIC_MSVC_BACKEND)
    so clang/clang-cl use the portable __atomic_* builtins instead of the
    __int64 MS extension token.

+++ app/filesystem_unix.c
@@ -199,7 +199,7 @@ static u64 get_file_id(const_chr_t path) {
 		return 0ull;
 	}
 	/* Combine device + inode into 64 bits (best-effort). */
-	return ((u64)st.st_dev << 32) ^ (u64)st.st_ino;
+	return ((u64)st.st_dev << 32) ^ st.st_ino;
 }
 
 static i32 create_directory(const_chr_t path) {
